+++ user/models/__init__.py
@@ -1,1 +1,1 @@
-from .modelos import *
+from .models import *
