+++ user/requirements.txt
@@ -1,4 +1,5 @@
-Flask==3.0.0
-SQLAlchemy
-Flask-SQLAlchemy
-flask-restful
+Flask==2.2.2
+Flask-RESTful==0.3.9
+Flask-SQLAlchemy==3.0.3
+SQLAlchemy==2.0.2
+Werkzeug==2.2.2
